list courses, Jombotrom and rating [for courses]

+++ database/migrations/2020_07_08_221813_create_reviews_table.php
@@ -17,7 +17,7 @@ class CreateReviewsTable extends Migration
             $table->increments('id');
             $table->unsignedInteger('course_id');
             $table->unsignedInteger('user_id');
-            $table->float('rating', 2);
+            $table->integer('rating');
             $table->mediumText('comment')->nullable();
             $table->timestamps();
